@@ -16,13 +16,13 @@
         }
     ],
     "require": {
-        "php": "^8.0"
+        "php": "^8.1|^8.2|^8.3"
     },
     "require-dev": {
         "laravel/pint": "^1.2",
-        "pestphp/pest": "^1.22",
+        "pestphp/pest": "^2",
         "phpstan/phpstan": "^1.6",
-        "rector/rector": "^0.15.0"
+        "rector/rector": "^1"
     },
     "autoload": {
         "psr-4": {
