@@ -2,43 +2,43 @@
 
 use HiFolks\DataType\Arr;
 
-it('is Array', function () {
+it('is Array', function (): void {
     $arr = Arr::make();
     expect($arr->arr())->toBeArray();
 });
 
-it('creates Arr from function', function () {
-    $arr = Arr::fromFunction(fn () => random_int(0, 100), 500);
+it('creates Arr from function', function (): void {
+    $arr = Arr::fromFunction(fn (): int => random_int(0, 100), 500);
     expect($arr->arr())->toBeArray();
     expect($arr->length())->toEqual(500);
 
-    expect($arr->every(fn ($element) => $element >= 0))->toBeTrue();
-    expect($arr->every(fn ($element) => $element <= 100))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element >= 0))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element <= 100))->toBeTrue();
 
     $arr = Arr::fromFunction(fn ($i) => $i, 5000);
     expect($arr->arr())->toBeArray();
     expect($arr->length())->toEqual(5000);
     //print_result($arr);
 
-    expect($arr->every(fn ($element) => $element >= 0))->toBeTrue();
-    expect($arr->every(fn ($element) => $element <= 5000))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element >= 0))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element <= 5000))->toBeTrue();
 });
 
-it('creates Arr from value', function () {
+it('creates Arr from value', function (): void {
     $arr = Arr::fromValue(0, 5000);
     expect($arr->arr())->toBeArray();
     expect($arr->length())->toEqual(5000);
-    expect($arr->every(fn ($element) => $element === 0))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element === 0))->toBeTrue();
 });
 
-it('is Array empty', function () {
+it('is Array empty', function (): void {
     $arr = Arr::make();
     expect($arr->arr())->toBeArray();
     expect(count($arr->arr()))->toEqual(0);
     expect($arr->count())->toEqual(0);
     expect($arr->length())->toEqual(0);
 });
-it('is Array length 1', function () {
+it('is Array length 1', function (): void {
     $arr = Arr::make([99]);
     expect($arr->arr())->toBeArray();
     expect(count($arr->arr()))->toEqual(1);
@@ -46,7 +46,7 @@ it('is Array length 1', function () {
     expect($arr->length())->toEqual(1);
 });
 
-it('is Array forEach', function () {
+it('is Array forEach', function (): void {
     $arr = Arr::make([99, 98]);
     expect($arr->arr())->toBeArray();
     expect(count($arr->arr()))->toEqual(2);
@@ -54,19 +54,19 @@ it('is Array forEach', function () {
     expect($arr->length())->toEqual(2);
     expect($arr->push(100))->toEqual(3);
     expect($arr->length())->toEqual(3);
-    $x = $arr->forEach(fn ($element, $key) => $key * $element.PHP_EOL);
+    $x = $arr->forEach(fn ($element, $key): string => $key * $element.PHP_EOL);
     expect($x->length())->toEqual(3);
     expect($x->get(2))->toEqual(200);
 });
 
-it('shift array', function () {
+it('shift array', function (): void {
     $arr = Arr::make([99, 98, 97]);
     expect($arr->length())->toEqual(3);
     expect($arr->shift())->toEqual(99);
     expect($arr->length())->toEqual(2);
 });
 
-it('unshift array', function () {
+it('unshift array', function (): void {
     $arr = Arr::make([99, 98, 97]);
     expect($arr->length())->toEqual(3);
     expect($arr->unshift(200))->toEqual(4);
@@ -79,7 +79,7 @@ it('unshift array', function () {
     $arr->unshift(-2, -1);
     expect($arr->toString())->toEqual('-2,-1,0,1,2');
 });
-it('append array', function () {
+it('append array', function (): void {
     $arr = Arr::make([99, 98, 97])->append([1, 2, 3]);
     expect($arr->length())->toEqual(6);
     $arr->append([12], [13, 14]);
@@ -91,19 +91,19 @@ it('append array', function () {
     expect($arr->length())->toEqual(0);
     expect($number)->toEqual(11);
 });
-it('joins arrays', function () {
+it('joins arrays', function (): void {
     $arr = Arr::make([99, 98, 97])->append([1, 2, 3]);
     expect($arr->join())->toEqual('99,98,97,1,2,3');
 });
 
-it('concats arrays', function () {
+it('concats arrays', function (): void {
     $arr = Arr::make([99, 98, 97])->append([1, 2, 3]);
     $arr2 = $arr->concat([1000, 1001]);
     expect($arr2->arr())->toBeArray();
     expect($arr2->length())->toEqual(8);
     expect($arr->length())->toEqual(6);
 });
-it('concats more types', function () {
+it('concats more types', function (): void {
     $arr = Arr::make([99, 98, 97])->concat([1, 2, 3], [1000, 1001]);
     expect($arr->arr())->toBeArray();
     expect($arr->length())->toEqual(8);
@@ -112,7 +112,7 @@ it('concats more types', function () {
     $arr3 = $arr->concat($arr2);
     expect($arr3->length())->toEqual(18);
 });
-it('concats indexed/associative arrays', function () {
+it('concats indexed/associative arrays', function (): void {
     $fruits = Arr::make([
         3 => '🥝',
         -1 => '🍓',
@@ -127,7 +127,7 @@ it('concats indexed/associative arrays', function () {
     expect($fruits2['mango'])->toEqual('🥭');
     expect($fruits2[4])->toEqual('🍍');
 });
-it('slices arrays', function () {
+it('slices arrays', function (): void {
     $arr = Arr::make([99, 98, 97])->append([1, 2, 3]);
     $arr2 = $arr->slice(1, 2);
     expect($arr2->arr())->toBeArray();
@@ -193,7 +193,7 @@ it('slices arrays', function () {
     expect($arr[0])->toBeNull();
 });
 
-it('searches arrays', function () {
+it('searches arrays', function (): void {
     $arr = Arr::make([99, 98, 97])->append([1, 2, 3]);
     $index = $arr->indexOf(1);
     expect($index)->toBeInt();
@@ -206,7 +206,7 @@ it('searches arrays', function () {
     expect($arr->length())->toEqual(3);
 });
 
-it('searches last occurrence arrays', function () {
+it('searches last occurrence arrays', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
     $index = $arr->lastIndexOf(5);
     expect($index)->toBeInt();
@@ -219,39 +219,39 @@ it('searches last occurrence arrays', function () {
     expect($arr->length())->toEqual(5);
 });
 
-it('matches every element', function () {
+it('matches every element', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    $bool = $arr->every(fn ($element) => $element > 0);
+    $bool = $arr->every(fn ($element): bool => $element > 0);
     expect($bool)->toEqual(true);
-    $bool = $arr->every(fn ($element) => $element > 1);
+    $bool = $arr->every(fn ($element): bool => $element > 1);
     expect($bool)->toEqual(false);
-    $bool = $arr->every(fn ($element) => $element > 10000);
+    $bool = $arr->every(fn ($element): bool => $element > 10000);
     expect($bool)->toEqual(false);
 });
 
-it('matches some element', function () {
+it('matches some element', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    $bool = $arr->some(fn ($element) => $element > 0);
+    $bool = $arr->some(fn ($element): bool => $element > 0);
     expect($bool)->toEqual(true);
-    $bool = $arr->some(fn ($element) => $element > 1);
+    $bool = $arr->some(fn ($element): bool => $element > 1);
     expect($bool)->toEqual(true);
-    $bool = $arr->some(fn ($element) => $element > 10000);
+    $bool = $arr->some(fn ($element): bool => $element > 10000);
     expect($bool)->toEqual(false);
 
     $arr = Arr::make([1, 2, 3, 4, 5]);
-    $even = fn ($element) => $element % 2 === 0;
+    $even = fn ($element): bool => $element % 2 === 0;
     $bool = $arr->some($even);
     expect($bool)->toEqual(true);
 });
 
-it('filters some element', function () {
+it('filters some element', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    $arr2 = $arr->filter(fn ($element) => $element > 3);
+    $arr2 = $arr->filter(fn ($element): bool => $element > 3);
     expect($arr->length())->toEqual(17);
     expect($arr2->length())->toEqual(11);
-    expect($arr->every(fn ($element) => $element > 0))->toBeTrue();
-    expect($arr->every(fn ($element) => $element <= 3))->toBeFalse();
-    expect($arr2->every(fn ($element) => $element > 3))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element > 0))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element <= 3))->toBeFalse();
+    expect($arr2->every(fn ($element): bool => $element > 3))->toBeTrue();
 });
 
 it('finds the first element')
@@ -261,48 +261,48 @@ it('finds the first element')
             'strArr' => Arr::make(['begin','middle','end'])
         ],
     ])
-    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element) => $element > 0))
+    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element): bool => $element > 0))
         ->toBeInt()->toEqual(1)
-    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element) => $element < 5))
+    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element): bool => $element < 5))
         ->toBeInt()->toEqual(1)
-    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element) => $element > 5))
+    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element): bool => $element > 5))
         ->toBeInt()->toEqual(6)
-    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element, $index) => $element > 1 && $index > 1))
+    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element, $index): bool => $element > 1 && $index > 1))
         ->toBeInt()->toEqual(3)
-    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element) => $element > 50))
+    ->expect(fn ($dataset) => $dataset->intArr->find(fn ($element): bool => $element > 50))
         ->toBeNull()
-    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element) => $element == 'end'))
+    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element): bool => $element == 'end'))
         ->toBeString()->toBe('end')
-    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element) => str_contains($element, 'e')))
+    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element): bool => str_contains((string) $element, 'e')))
         ->toBeString()->toBe('begin')
-    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element) => $element == 'prefix'))
+    ->expect(fn ($dataset) => $dataset->strArr->find(fn ($element): bool => $element == 'prefix'))
         ->toBeNull();
 
-it('finds the first index of some element', function () {
+it('finds the first index of some element', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    $index = $arr->findIndex(fn ($element) => $element > 0);
+    $index = $arr->findIndex(fn ($element): bool => $element > 0);
     expect($index)->toBeInt();
     expect($index)->toEqual(0);
-    $index = $arr->findIndex(fn ($element) => $element > 1);
+    $index = $arr->findIndex(fn ($element): bool => $element > 1);
     expect($index)->toEqual(1);
-    $index = $arr->findIndex(fn ($element) => $element > 10000);
+    $index = $arr->findIndex(fn ($element): bool => $element > 10000);
     expect($index)->toEqual(-1);
-    $index = $arr->findIndex(fn ($element, $index) => $element > 1 && $index > 1);
+    $index = $arr->findIndex(fn ($element, $index): bool => $element > 1 && $index > 1);
     expect($index)->toEqual(2);
 });
 
-it('maps elements', function () {
+it('maps elements', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    $arr2 = $arr->map(fn ($element) => $element + 1);
+    $arr2 = $arr->map(fn ($element): int|float => $element + 1);
 
     expect($arr->length())->toEqual(17);
     expect($arr2->length())->toEqual(17);
-    expect($arr->every(fn ($element) => $element > 0))->toBeTrue();
-    expect($arr->every(fn ($element) => $element > 1))->toBeFalse();
-    expect($arr2->every(fn ($element) => $element > 1))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element > 0))->toBeTrue();
+    expect($arr->every(fn ($element): bool => $element > 1))->toBeFalse();
+    expect($arr2->every(fn ($element): bool => $element > 1))->toBeTrue();
 });
 
-it('flats array', function () {
+it('flats array', function (): void {
     $arr = Arr::make([1, [2, 3], 4, [5, 6, 7]]);
 
     $arr2 = $arr->flat();
@@ -310,10 +310,10 @@ it('flats array', function () {
     expect($arr2->length())->toEqual(7);
 });
 
-it('flats and maps array', function () {
+it('flats and maps array', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
 
-    $arr2 = $arr->flatMap(fn ($element) => [$element, $element * 2]);
+    $arr2 = $arr->flatMap(fn ($element): array => [$element, $element * 2]);
     expect($arr->length())->toEqual(7);
     expect($arr2->length())->toEqual(14);
     expect($arr2->arr())->toEqual([1, 2, 2, 4, 3, 6, 4, 8, 5, 10, 6, 12, 7, 14]);
@@ -324,7 +324,7 @@ it('flats and maps array', function () {
     expect($arr2->arr())->toEqual([1, 2, 3, 4, 5, 6, 7]);
 });
 
-it('fills array', function () {
+it('fills array', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
     expect($arr->length())->toEqual(7);
     $arr->fill(0, 0, 3);
@@ -336,30 +336,30 @@ it('fills array', function () {
     expect($arr[4])->toEqual(5);
     expect($arr[5])->toEqual(6);
     expect($arr[6])->toEqual(7);
-    $arr2 = $arr->filter(fn ($element) => $element == 0);
+    $arr2 = $arr->filter(fn ($element): bool => $element == 0);
     expect($arr2->length())->toEqual(4);
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
     $arr->fill(0);
-    $allAreZeros = $arr->every(fn ($element) => $element === 0);
+    $allAreZeros = $arr->every(fn ($element): bool => $element === 0);
     expect($arr->length())->toEqual(7);
     expect($allAreZeros)->toBeTrue();
 });
 
-it('reduces Arr', function () {
+it('reduces Arr', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
-    $value = $arr->reduce(fn ($previousValue, $currentValue) => $previousValue + $currentValue);
+    $value = $arr->reduce(fn ($previousValue, $currentValue): float|int|array => $previousValue + $currentValue);
     expect($value)->toBeInt();
     expect($value)->toEqual(28);
 });
 
-it('reduces Arr in reverse way', function () {
+it('reduces Arr in reverse way', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
-    $value = $arr->reduceRight(fn ($previousValue, $currentValue) => $previousValue + $currentValue);
+    $value = $arr->reduceRight(fn ($previousValue, $currentValue): float|int|array => $previousValue + $currentValue);
     expect($value)->toBeInt();
     expect($value)->toEqual(28);
 });
 
-it('reverses Arr', function () {
+it('reverses Arr', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7]);
     $arr2 = $arr->reverse();
     expect($arr2->length())->toEqual(7);
@@ -370,14 +370,14 @@ it('reverses Arr', function () {
     expect($arr->join())->toEqual('7,6,5,4,3,2,1,0');
 });
 
-it('sorts Arr', function () {
+it('sorts Arr', function (): void {
     $arr = Arr::make([6, 2, 4, 2, 1, 9, 7]);
     expect($arr->join())->toEqual('6,2,4,2,1,9,7');
     $arr->sort();
     expect($arr->length())->toEqual(7);
     expect($arr->join())->toEqual('1,2,2,4,6,7,9');
 });
-it('sorts and change Arr', function () {
+it('sorts and change Arr', function (): void {
     $months = Arr::make(['March', 'Jan', 'Feb', 'Dec']);
     $monthsSorted = $months->sort();
     expect($months->length())->toEqual(4);
@@ -390,7 +390,7 @@ it('sorts and change Arr', function () {
     expect($monthsSorted[3])->toBeNull();
 });
 
-it('splices Arr', function () {
+it('splices Arr', function (): void {
     $months = Arr::make(['Jan', 'March', 'April', 'June']);
     $months->splice(1, 0, 'Feb');
     expect($months->join())->toEqual('Jan,Feb,March,April,June');
@@ -400,7 +400,7 @@ it('splices Arr', function () {
     expect($months->join())->toEqual('Jan');
 });
 
-it('stringifies an Arr', function () {
+it('stringifies an Arr', function (): void {
     $months = Arr::make(['Jan', 'Feb', 'March', 'April', 'May']);
     expect($months->toString())->toEqual('Jan,Feb,March,April,May');
 
@@ -408,7 +408,7 @@ it('stringifies an Arr', function () {
     expect($arr->toString())->toEqual('1,2,a,1a');
 });
 
-it('checks is array', function () {
+it('checks is array', function (): void {
     $isArray = Arr::isArray(['Jan', 'Feb', 'March', 'April', 'May']);
     expect($isArray)->toEqual(true);
     $isArray = Arr::isArray(null);
@@ -419,20 +419,20 @@ it('checks is array', function () {
     expect($isArray)->toEqual(false);
 });
 
-it('tests if array is empty', function () {
+it('tests if array is empty', function (): void {
     $emptyArr = Arr::make([]);
     expect($emptyArr->isEmpty())->toEqual(true);
     $notEmptyArr = Arr::make([1]);
     expect($notEmptyArr->isEmpty())->toEqual(false);
 });
 
-it('chainable empty array', function () {
+it('chainable empty array', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
-    expect($arr->filter(fn ($element) => $element > 100)->isEmpty())->toBeTrue();
-    expect($arr->filter(fn ($element) => $element < 100)->isEmpty())->toBeFalse();
+    expect($arr->filter(fn ($element): bool => $element > 100)->isEmpty())->toBeTrue();
+    expect($arr->filter(fn ($element): bool => $element < 100)->isEmpty())->toBeFalse();
 });
 
-it('implements of() method', function () {
+it('implements of() method', function (): void {
     $arr = Arr::of('Jan', 'Feb', 'March', 'April', 'May');
     expect($arr->length())->toEqual(5);
     expect($arr[4])->toEqual('May');
@@ -442,7 +442,7 @@ it('implements of() method', function () {
     expect($arr[0])->toEqual(7);
 });
 
-it('tests keys() method', function () {
+it('tests keys() method', function (): void {
     $arr = Arr::make(
         [
             '01' => 'Jan', '02' => 'Feb', '03' => 'March', '04' => 'April',
@@ -466,7 +466,7 @@ it('tests keys() method', function () {
     }
 });
 
-it('test keys() method', function () {
+it('test keys() method', function (): void {
     $arr = Arr::make(
         [5, 12, 8, 130, 44]
     );
@@ -476,7 +476,7 @@ it('test keys() method', function () {
     expect($arr->at(100))->toBeNull();
 });
 
-it('includes element', function () {
+it('includes element', function (): void {
     $arr = Arr::make(
         [1, 2, 3]
     );
@@ -493,7 +493,7 @@ it('includes element', function () {
     expect($arr->includes('at'))->toBeFalse();
 });
 
-it('includes element from index', function () {
+it('includes element from index', function (): void {
     $arr = Arr::make(
         [1, 2, 3]
     );
@@ -516,7 +516,7 @@ it('includes element from index', function () {
     expect($arr->includes('a', -2))->toBeFalse();
 });
 
-it(' extract values', function () {
+it(' extract values', function (): void {
     $fruits = Arr::make([
         7 => '🥝',
         -1 => '🍓',
@@ -544,7 +544,7 @@ it(' extract values', function () {
     expect($i)->toEqual(8);
 });
 
-it('creates entries', function () {
+it('creates entries', function (): void {
     $fruits = Arr::make([
         7 => '🥝',
         -1 => '🍓',
@@ -571,7 +571,7 @@ it('creates entries', function () {
     ]));
 });
 
-it('tests copyWithin() method with one parameter', function () {
+it('tests copyWithin() method with one parameter', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5]);
 
     $result = $arr->copyWithin(-2);
@@ -581,7 +581,7 @@ it('tests copyWithin() method with one parameter', function () {
         ->toEqual([1, 2, 3, 1, 2]);
 });
 
-it('tests copyWithin() method with two parameters', function () {
+it('tests copyWithin() method with two parameters', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5]);
 
     $result = $arr->copyWithin(0, 3);
@@ -591,7 +591,7 @@ it('tests copyWithin() method with two parameters', function () {
         ->toEqual([4, 5, 3, 4, 5]);
 });
 
-it('tests copyWithin() method with all the parameters', function () {
+it('tests copyWithin() method with all the parameters', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5]);
 
     $result = $arr->copyWithin(0, 3, 4);
@@ -601,7 +601,7 @@ it('tests copyWithin() method with all the parameters', function () {
         ->toEqual([4, 2, 3, 4, 5]);
 });
 
-it('tests copyWithin() method with all the parameters negative', function () {
+it('tests copyWithin() method with all the parameters negative', function (): void {
     $arr = Arr::make([1, 2, 3, 4, 5]);
 
     $result = $arr->copyWithin(-2, -3, -1);
@@ -611,18 +611,18 @@ it('tests copyWithin() method with all the parameters negative', function () {
         ->toEqual([1, 2, 3, 3, 4]);
 });
 
-it('tests flatMap can handle an array of arrays', function () {
+it('tests flatMap can handle an array of arrays', function (): void {
     $arr = Arr::make([
         [1, 2],
         [3, 4]
     ]);
 
-    $result = $arr->flatMap(fn ($element) => $element * 2);
+    $result = $arr->flatMap(fn ($element): int|float => $element * 2);
     expect($result->length())->toEqual(4);
     expect($result->arr())->toEqual([2, 4, 6, 8]);
 });
 
-it('can unset array elements by their keys', function () {
+it('can unset array elements by their keys', function (): void {
     $arr = Arr::make([
         'mango' => '🥭',
         'apple' => '🍎',
@@ -648,7 +648,7 @@ it('can unset array elements by their keys', function () {
         ]);
 });
 
-it('can set array key', function () {
+it('can set array key', function (): void {
     $arr = Arr::make([
         'mango' => '🥭',
         'banana' => '🍌'
