@@ -8,11 +8,11 @@ class Operation
 {
     public static function add(int|string $field, int|float $value): Closure
     {
-        return fn ($element) => $value + $element[$field];
+        return fn ($element): int|float => $value + $element[$field];
     }
 
     public static function double(int|string $field): Closure
     {
-        return fn ($element) => $element[$field] * 2;
+        return fn ($element): int|float => $element[$field] * 2;
     }
 }
